Pipeline rework
Move to a model where the pipeline core is fixed, and we can plug in
photo and video bins to the tee element, and/or add an audiobin
separated from the rest for recording audio. Fixes problems caused by
not doing pad blocking and having unconnected elements lying around
in the pipeline.

Video recording is not quite working right... Audio and photo seem OK.