Patch to resolve trac #8580.
Record was unmapping its window prior to close via a call to self.hide(), which
resolves to gtk.Window.hide().  Then we attempted to capture a screenshot per
trac #8432.  This triggered a BadMatch X windows error when we tried to take a
screenshot of it using XShmGetImage (in sugar-preview.c in _sugarext in
sugar-toolkit).

This commit simply removes the hide() calls in Record. This prevents the window
from being unmapped and avoids this error.