Pipeline rework
Record-59 is broken in joyride because we moved to a gstreamer, which
doesn't like these broken pipelines.

This commit reworks the whole pipeline system, making it easier to
read and generally better structured.

This is the work I did this summer, which didn't quite work for 8.2.
Now that we have upgraded to the newer gstreamer in F10, the bugs
have disappeared. There are a couple of quirks with video recording
but it works well.