Record photo UUID as string
Photos were not being shared properly (the download never succeeded)
because the servers MD5 checking always failed ("I don't have this photo").

This is because the UUIDs were always being saved as UUID objects, and
comparing them later with a string (in getRecdByMd5) was always failing.

Save UUIDs as a string to solve this issue and others that surround it.