Use icons for timer/duration selection (SL#2972)
These strings were too long; in Spanish the Stop button was pushed
off the end of the toolbar.

Replace them with icons for compactness, based on input from the design
team.

It would be nicer to have these icons only showing in the unexpanded
rendering of the comboboxes, but this does not seem possible:
http://lists.sugarlabs.org/archive/sugar-devel/2011-July/032363.html