Don't use pad blocking
This never worked right: we always managed to block, but the unblock
never happened.

Previously this didn't hurt anything, but on F17 it causes problems: since
the pad is blocked, we never take the photo.

Remove pad blocking to work around this. Seems to be working fine
with live-pipeline modification.