Set Analog Mic Boost control on XO-{1.75,4}
Set controls for best voice recording in quiet environment.

Audio recording level on XO-1.75 and XO-4 is set by two ALSA mixer
controls, Capture and Analog Mic Boost.

Capture is exposed on the Sugar Frame as Microphone.  In noisy
places learners can reduce the Capture control to fix clipping.

Analog Mic Boost is not exposed in Sugar, is set to zero by Measure
activity quit, and persists at zero over reboots thanks to alsactl.

(Record might not capture any sound, or might only capture tapping on
the microphone.  But this was not easily reproduced.  The key discovery
is that testers often test alphabetically, and with the Measure activity
run first the Record activity will fail to capture sound.)

On XO-1.75 requires a new kernel.  On XO-4 works immediately.

http://dev.laptop.org/ticket/12867

Signed-off-by: James Cameron <quozl@laptop.org>