Support wider aspect ratios
Displays with wider aspect ratios of 16:9 gave squashed thumbnails and
photograph views, because the activity assumed a 4:3 aspect ratio.

When a photograph was taken, the thumbnail image was shown squashed,
people were tall and thin.  When a photograph was selected from the
tray, the same effect was shown in the media view.  The live camera view
was also smaller than it could be.

Fix is to;

- detect the display aspect ratio, and use it for sizing the media view
  widget, and;

- detect the camera aspect ratio, and trim sides of image to fit our 4:3
  thumbnail SVG template.

Fixes most but not all of https://bugs.sugarlabs.org/ticket/4860