Fix copy to clipboard
Previously, Record would send local filesystem URIs in the clipboard for
all three modes.  Nothing else seems to work with this; Frame ignores
it, Paint ignores it, and Jukebox ignores it.

Switch to sending only an image.  For the Photo and Audio mode, a camera
resolution image was already available.  For the Video mode, add this.

Did not add to collaboration, as this code is not yet tested after port
to GTK+ 3, and there's a really ugly hack used.