Fix fullscreen toggle of live view on Ubuntu 16.04
On Ubuntu 16.04 with GStreamer 1.8, the fullscreen toggle of live view
gave black display.

Cause was a dangling reference to last-pixbuf property of the
gdkpixbufsink, which prevented the pipeline from terminating, in turn
leaving the camera device open.  When the stream was restarted, the
camera device was in use.

There were three potential references to the last-pixbuf;

- self._pixbuf, now destroyed when pipeline is stopped,

- self._audio_pixbuf, now made a data copy,

- self._video_pixbuf, now made a data copy.

Test case; ctrl-f shortcut repeatedly.